Added filtered functionality

+++ test.sh
@@ -47,7 +47,7 @@ append_to_csv_file() {
 delete_entry_by_id() {
     filename=$1
     read -p "Enter id:" id
-    if [-f "$filename"]; then
+    if [ -f "$filename" ]; then
         grep -v -w "$id" "$filename" > temp.txt && mv temp.txt "$filename"
     else 
         echo "File $filename does not exits."
@@ -55,29 +55,42 @@ delete_entry_by_id() {
 }
 edit_row() {
     filename=$1
-    regex="^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$"
-    read -p "Enter id: " id
-    read -p "Enter name: " name
-    while true; do 
-        read -p "Enter notaSO: " notaSO
-        if [[ "$notaSO" =~ ^[1-9]$ ]] || [[ "$notaSO" == 10 ]]; then
-            break
-        else 
-            echo "Invalid number. Please enter a number between 1 and 10"
-        fi
-    done
-    while true; do
-        read -p "Enter email: " email
-        if [[ $email =~ $regex ]]; then
-            break
-        else 
-            echo "Invalid email. Please try again!"
-        fi
-    done
-    local new_row="$id,$name,$notaSO,$email"
-    local temp_file=$(mktemp)
+    if [ -f "$filename" ]; then
+        regex="^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$"
+        read -p "Enter id: " id
+        read -p "Enter name: " name
+        while true; do 
+            read -p "Enter notaSO: " notaSO
+            if [[ "$notaSO" =~ ^[1-9]$ ]] || [[ "$notaSO" == 10 ]]; then
+                break
+            else 
+                echo "Invalid number. Please enter a number between 1 and 10"
+            fi
+        done
+        while true; do
+            read -p "Enter email: " email
+            if [[ $email =~ $regex ]]; then
+                break
+            else 
+                echo "Invalid email. Please try again!"
+            fi
+        done
+        local new_row="$id,$name,$notaSO,$email"
+        local temp_file=$(mktemp)
 
-    awk -v id="$id" -v row="$new_row" 'BEGIN{FS=OFS=","} $1 != id {print $0} $1 == id {print row}' "$filename" > "$temp_file" && mv "$temp_file" "$filename"
+        awk -v id="$id" -v row="$new_row" 'BEGIN{FS=OFS=","} $1 != id {print $0} $1 == id {print row}' "$filename" > "$temp_file" && mv "$temp_file" "$filename"
+    else 
+        echo "File $filename does not exist."
+    fi
+}
+
+sort_by_notaSO() {
+    filename=$1
+    if [ -f "$filename" ]; then
+        sort -t ',' -k3,3nr "$filename" | head -n 4
+    else
+        echo "File $filename does not exist."
+    fi
 }
 # Main loop
 while true; do
@@ -86,6 +99,7 @@ while true; do
     echo "2. Edit CSV row by id"
     echo "3. Append content to CSV file"
     echo "4. Delete entry of CSV file"
+    echo "5. Sort by notaSO"
     echo "0. Exit"
 
     read -p "Enter your choice: " choice
@@ -106,6 +120,10 @@ while true; do
         4)
             read -p "Enter the filename: " filename
             delete_entry_by_id "$filename"
+            ;;
+        5)
+            read -p "Enter the filename: " filename
+            sort_by_notaSO "$filename"
             ;;
         0)
             echo "Exiting..."
